code/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_ipyw_jlab_nb_ext_conf=0.1.0=py38_0
abseil-cpp=20230802.0=h5da7b33_2
absl-py=1.4.0=py38haa95532_0
aiohttp=3.9.0=py38h2bbff1b_0
aiosignal=1.2.0=pyhd3eb1b0_0
anaconda-anon-usage=0.4.3=py38h8ddfbff_0
anaconda-client=1.12.2=py38haa95532_0
anaconda-navigator=1.10.0=py38_0
anyio=3.5.0=py38haa95532_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py38h2bbff1b_0
asttokens=2.0.5=pyhd3eb1b0_0
async-lru=2.0.4=py38haa95532_0
async-timeout=4.0.3=py38haa95532_0
attrs=23.1.0=py38haa95532_0
babel=2.11.0=py38haa95532_0
backcall=0.2.0=pyhd3eb1b0_0
backports=1.1=pyhd3eb1b0_0
backports.functools_lru_cache=1.6.4=pyhd3eb1b0_0
backports.tempfile=1.0=pyhd3eb1b0_1
backports.weakref=1.0.post1=py_1
beautifulsoup4=4.12.2=py38haa95532_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
blinker=1.6.2=py38haa95532_0
boltons=23.0.0=py38haa95532_0
brotli-python=1.0.9=py38hd77b12b_7
bzip2=1.0.8=he774522_0
c-ares=1.19.1=h2bbff1b_0
ca-certificates=2023.12.12=haa95532_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2023.11.17=py38haa95532_0
cffi=1.16.0=py38h2bbff1b_0
chardet=4.0.0=py38haa95532_1003
charset-normalizer=2.0.4=pyhd3eb1b0_0
clang=5.0=pypi_0
click=8.1.7=py38haa95532_0
colorama=0.4.6=py38haa95532_0
comm=0.1.2=py38haa95532_0
conda=23.5.0=py38haa95532_0
conda-build=3.20.5=py38_1
conda-env=2.6.0=1
conda-package-handling=2.2.0=py38haa95532_0
conda-package-streaming=0.9.0=py38haa95532_0
conda-verify=3.4.2=py_1
console_shortcut=0.1.1=4
cryptography=41.0.7=py38h89fc84f_0
cycler=0.11.0=pyhd3eb1b0_0
debugpy=1.6.7=py38hd77b12b_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
docker-pycreds=0.4.0=pypi_0
dominate=2.7.0=pypi_0
executing=0.8.3=pyhd3eb1b0_0
filelock=3.13.1=py38haa95532_0
flatbuffers=1.12=pypi_0
freetype=2.12.1=ha860e81_0
frozenlist=1.4.0=py38h2bbff1b_0
future=0.18.3=py38haa95532_0
giflib=5.2.1=h8cc25b3_3
gitdb=4.0.9=pypi_0
gitpython=3.1.27=pypi_0
glob2=0.7=pyhd3eb1b0_0
google-auth=2.22.0=py38haa95532_0
google-auth-oauthlib=0.4.4=pyhd3eb1b0_0
grpc-cpp=1.48.2=h6772dbd_4
grpcio=1.41.0=pypi_0
gtest=1.14.0=h59b6b97_0
h5py=3.1.0=pypi_0
hdf5=1.10.4=h7ebc959_0
icc_rt=2022.1.0=h6049295_2
icu=73.1=h6c2663c_0
idna=3.4=py38haa95532_0
importlib-metadata=7.0.1=py38haa95532_0
importlib_metadata=7.0.1=hd3eb1b0_0
importlib_resources=6.1.1=py38haa95532_1
intel-openmp=2021.4.0=haa95532_3556
ipykernel=6.25.0=py38h9909e9c_0
ipython=8.12.2=py38haa95532_0
ipywidgets=8.0.4=py38haa95532_0
jedi=0.18.1=py38haa95532_1
jinja2=3.1.2=py38haa95532_0
jpeg=9e=h2bbff1b_1
json5=0.9.6=pyhd3eb1b0_0
jsonpatch=1.32=pyhd3eb1b0_0
jsonpointer=2.3=pypi_0
jsonschema=4.19.2=py38haa95532_0
jsonschema-specifications=2023.7.1=py38haa95532_0
jupyter-lsp=2.2.0=py38haa95532_0
jupyter_client=8.6.0=py38haa95532_0
jupyter_core=5.5.0=py38haa95532_0
jupyter_events=0.8.0=py38haa95532_0
jupyter_server=2.10.0=py38haa95532_0
jupyter_server_terminals=0.4.4=py38haa95532_1
jupyterlab=4.0.8=py38haa95532_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.25.1=py38haa95532_0
jupyterlab_widgets=3.0.9=py38haa95532_0
keras=2.6.0=pypi_0
kiwisolver=1.4.4=py38hd77b12b_0
krb5=1.20.1=h5b6d351_0
lerc=3.0=hd77b12b_0
libarchive=3.6.2=hb62f4d4_2
libclang=14.0.6=default_hb5a9fac_1
libclang13=14.0.6=default_h8e68704_1
libdeflate=1.17=h2bbff1b_1
libffi=3.4.4=hd77b12b_0
libiconv=1.16=h2bbff1b_2
liblief=0.12.3=hd77b12b_0
libopencv=4.0.1=hbb9e17c_0
libpng=1.6.39=h8cc25b3_0
libpq=12.15=h906ac69_1
libprotobuf=3.20.3=h23ce68f_0
libsodium=1.0.18=h62dcd97_0
libtiff=4.5.1=hd77b12b_0
libuv=1.44.2=h2bbff1b_0
libwebp=1.3.2=hbc33d0d_0
libwebp-base=1.3.2=h2bbff1b_0
libxml2=2.10.4=h0ad7f3c_1
lz4-c=1.9.4=h2bbff1b_0
markdown=3.4.1=py38haa95532_0
markupsafe=2.1.3=py38h2bbff1b_0
matplotlib=3.3.4=py38haa95532_0
matplotlib-base=3.3.4=py38h49ac443_0
matplotlib-inline=0.1.6=py38haa95532_0
menuinst=1.4.16=py38he774522_1
mistune=2.0.4=py38haa95532_0
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py38h2bbff1b_0
mkl_fft=1.3.1=py38h277e83a_0
mkl_random=1.2.2=py38hf11a4ad_0
multidict=6.0.4=py38h2bbff1b_0
navigator-updater=0.2.1=py38_0
nbclient=0.8.0=py38haa95532_0
nbconvert=7.10.0=py38haa95532_0
nbformat=5.9.2=py38haa95532_0
nest-asyncio=1.5.6=py38haa95532_0
ninja=1.10.2=haa95532_5
ninja-base=1.10.2=h6d14046_5
notebook-shim=0.2.3=py38haa95532_0
numpy=1.24.3=py38hf95b240_0
numpy-base=1.24.3=py38h005ec55_0
oauthlib=3.2.2=py38haa95532_0
opencv=4.0.1=py38h2a7c758_0
opencv-python=4.5.2.54=pypi_0
openjpeg=2.4.0=h4fc8c34_0
openssl=3.0.12=h2bbff1b_0
overrides=7.4.0=py38haa95532_0
packaging=23.1=py38haa95532_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=10.0.1=py38h045eedc_0
pip=21.0.1=pypi_0
pkginfo=1.9.6=py38haa95532_0
pkgutil-resolve-name=1.3.10=py38haa95532_1
platformdirs=3.10.0=py38haa95532_0
pluggy=1.0.0=py38haa95532_1
ply=3.11=py38_0
powershell_shortcut=0.0.1=3
prometheus_client=0.14.1=py38haa95532_0
promise=2.3=pypi_0
prompt-toolkit=3.0.43=py38haa95532_0
protobuf=3.20.3=py38hd77b12b_0
psutil=5.9.0=py38h2bbff1b_0
pure_eval=0.2.2=pyhd3eb1b0_0
py-lief=0.12.3=py38hd77b12b_0
py-opencv=4.0.1=py38he44ac1e_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycosat=0.6.6=py38h2bbff1b_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.15.1=py38haa95532_1
pyjwt=2.4.0=py38haa95532_0
pyopenssl=23.2.0=py38haa95532_0
pyparsing=3.0.9=py38haa95532_0
pyqt=5.15.10=py38hd77b12b_0
pyqt5-sip=12.13.0=py38h2bbff1b_0
pysocks=1.7.1=py38haa95532_0
python=3.8.18=h1aa4202_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.16.2=py38haa95532_0
python-json-logger=2.0.7=py38haa95532_0
python-libarchive-c=2.9=pyhd3eb1b0_1
pytorch=1.12.1=cpu_py38h5e1f01c_0
pytz=2023.3.post1=py38haa95532_0
pywin32=305=py38h2bbff1b_0
pywinpty=2.0.10=py38h5da7b33_0
pyyaml=6.0.1=py38h2bbff1b_0
pyzmq=25.1.0=py38hd77b12b_0
qt-main=5.15.2=h19c9488_10
qtpy=2.4.1=py38haa95532_0
re2=2022.04.01=hd77b12b_0
referencing=0.30.2=py38haa95532_0
requests=2.31.0=py38haa95532_0
requests-oauthlib=1.3.0=py_0
requests-toolbelt=1.0.0=py38haa95532_0
rfc3339-validator=0.1.4=py38haa95532_0
rfc3986-validator=0.1.1=py38haa95532_0
rpds-py=0.10.6=py38h062c2fa_0
rsa=4.7.2=pyhd3eb1b0_1
ruamel.yaml=0.17.21=py38h2bbff1b_0
ruamel.yaml.clib=0.2.6=py38h2bbff1b_1
send2trash=1.8.2=py38haa95532_0
sentry-sdk=1.9.5=pypi_0
setproctitle=1.3.2=pypi_0
setuptools=68.2.2=py38haa95532_0
shortuuid=1.0.9=pypi_0
sip=6.7.12=py38hd77b12b_0
six=1.15.0=pypi_0
smmap=5.0.0=pypi_0
sniffio=1.2.0=py38haa95532_1
soupsieve=2.5=py38haa95532_0
sqlite=3.41.2=h2bbff1b_0
stack_data=0.2.0=pyhd3eb1b0_0
tensorboard=2.7.0=pypi_0
tensorboard-data-server=0.6.1=py38haa95532_0
tensorboard-plugin-wit=1.8.1=py38haa95532_0
tensorboardx=2.4=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.17.1=py38haa95532_0
tinycss2=1.2.1=py38haa95532_0
tk=8.6.12=h2bbff1b_0
tomli=2.0.1=py38haa95532_0
toolz=0.12.0=py38haa95532_0
torch=1.9.0=pypi_0
torchfile=0.1.0=pypi_0
torchvision=0.5.0=pypi_0
tornado=6.3.3=py38h2bbff1b_0
tqdm=4.65.0=py38hd4e2768_0
traitlets=5.7.1=py38haa95532_0
typing-extensions=3.7.4.3=pypi_0
typing_extensions=4.7.1=py38haa95532_0
urllib3=1.26.12=pypi_0
vc=14.2=h21ff451_1
visdom=0.1.8.9=pypi_0
vs2015_runtime=14.27.29016=h5e58377_2
wandb=0.13.2=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py38_1
websocket-client=1.4.0=pypi_0
werkzeug=2.2.3=py38haa95532_0
wheel=0.41.2=py38haa95532_0
widgetsnbextension=4.0.5=py38haa95532_0
win_inet_pton=1.1.0=py38haa95532_0
winpty=0.4.3=4
xmltodict=0.13.0=py38haa95532_0
xz=5.4.5=h8cc25b3_0
yaml=0.2.5=he774522_0
yarl=1.9.3=py38h2bbff1b_0
zeromq=4.3.4=hd77b12b_0
zipp=3.17.0=py38haa95532_0
zlib=1.2.13=h8cc25b3_0
zstandard=0.19.0=py38h2bbff1b_0
zstd=1.5.5=hd43e919_0
